Roccat Vulcan 100/120 initial driver
A userspace driver can be used to control the button LEDs of the
keyboard. As soon as the initialization sequence is sent, the
keyboard changes what data is sent for the media keys, and additional
data is sent for every key press and release which can be used to detect
key events previously only handled by the keyboard itself.

This driver reports special keys that the generic HID driver misses
when using the userspace driver, including:
- Profile keys (FN_F1 - FN_F4)
- Media keys (play-pause, stop, next song, previous song)
- FN key

Signed-off-by: Tobias Büttner <t.linux@spiritcroc.de>